simplify AddExpenseForm & index, & move css to Sections
